refactor: RacingCarGame.js 파일 분리
